backend-build/requirements-backend.txt
# Lightweight backend requirements (ML models run in separate microservices)

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0

# Core utilities
pandas
numpy
requests
tqdm
httpx>=0.25.0

# PDF/Document processing (lightweight)
pdfplumber
pypdfium2

# Text processing
markdown
beautifulsoup4
lxml

# Data processing
scikit-learn
pillow

# FAISS for vector search (CPU only, no torch dependency)
faiss-cpu

# LLM client
ollama
tiktoken

# Environment
python-dotenv

# PostgreSQL async support
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
sqlalchemy[asyncio]>=2.0.23
greenlet>=3.0.0
argon2_cffi>=21.0.0
